Refactor champion cog and add permission decorator
