+++ bot.py
@@ -158,6 +158,10 @@ class MyBot(commands.Bot):
         await self.tree.sync(guild=self.main_guild)
 
     async def on_ready(self):
+        if not isinstance(self.main_guild, discord.Guild):
+            guild = self.get_guild(self.main_guild_id)
+            if guild:
+                self.main_guild = guild
         logger.info(
             f"Bot ist bereit! Eingeloggt als {self.user} (ID: {self.user.id})")
