+++ cogs/champion/slash_commands.py
@@ -2,6 +2,7 @@ import discord
 from discord import app_commands
 
 from log_setup import get_logger
+from permissions import moderator_only
 from .cog import ChampionCog
 
 logger = get_logger(__name__)
@@ -13,11 +14,9 @@ champion_group = app_commands.Group(
 
 
 @champion_group.command(name="give", description="Gibt einem User Punkte (nur Mods)")
+@moderator_only()
 @app_commands.describe(user="Der Nutzer, dem Punkte gegeben werden", punkte="Anzahl der Punkte", grund="Begründung")
 async def give(interaction: discord.Interaction, user: discord.Member, punkte: int, grund: str):
-    if not interaction.user.guild_permissions.manage_guild:
-        await interaction.response.send_message("❌ Du hast keine Berechtigung für diesen Befehl.", ephemeral=True)
-        return
 
     logger.info(
         f"/champion give by {interaction.user} -> {user} ({punkte} Punkte, Grund: {grund})"
@@ -29,11 +28,9 @@ async def give(interaction: discord.Interaction, user: discord.Member, punkte: i
 
 
 @champion_group.command(name="remove", description="Entfernt Punkte (nur Mods)")
+@moderator_only()
 @app_commands.describe(user="Der Nutzer, von dem Punkte abgezogen werden", punkte="Anzahl der Punkte", grund="Begründung")
 async def remove(interaction: discord.Interaction, user: discord.Member, punkte: int, grund: str):
-    if not interaction.user.guild_permissions.manage_guild:
-        await interaction.response.send_message("❌ Du hast keine Berechtigung für diesen Befehl.", ephemeral=True)
-        return
 
     logger.info(
         f"/champion remove by {interaction.user} -> {user} ({punkte} Punkte, Grund: {grund})"
@@ -45,11 +42,9 @@ async def remove(interaction: discord.Interaction, user: discord.Member, punkte:
 
 
 @champion_group.command(name="set", description="Setzt die Punktzahl eines Users (nur Mods)")
+@moderator_only()
 @app_commands.describe(user="Der Nutzer, dessen Punktzahl gesetzt wird", punkte="Neue Gesamtpunktzahl", grund="Begründung")
 async def set_points(interaction: discord.Interaction, user: discord.Member, punkte: int, grund: str):
-    if not interaction.user.guild_permissions.manage_guild:
-        await interaction.response.send_message("❌ Du hast keine Berechtigung für diesen Befehl.", ephemeral=True)
-        return
 
     logger.info(
         f"/champion set by {interaction.user} -> {user} ({punkte} Punkte, Grund: {grund})"
@@ -63,11 +58,9 @@ async def set_points(interaction: discord.Interaction, user: discord.Member, pun
 
 
 @champion_group.command(name="reset", description="Setzt die Punkte eines Nutzers auf 0 (nur Mods)")
+@moderator_only()
 @app_commands.describe(user="Der Nutzer, dessen Punkte zurückgesetzt werden")
 async def reset(interaction: discord.Interaction, user: discord.Member):
-    if not interaction.user.guild_permissions.manage_guild:
-        await interaction.response.send_message("❌ Du hast keine Berechtigung für diesen Befehl.", ephemeral=True)
-        return
 
     logger.info(f"/champion reset by {interaction.user} for {user}")
 
@@ -118,12 +111,10 @@ async def myhistory(interaction: discord.Interaction):
 
 
 @champion_group.command(name="history", description="Zeigt die Punkte-Historie eines Spielers")
+@moderator_only()
 @app_commands.describe(user="Der Spieler, dessen Historie angezeigt wird")
 async def history(interaction: discord.Interaction, user: discord.Member):
     logger.info(f"/champion history by {interaction.user} target={user}")
-    if not interaction.user.guild_permissions.manage_guild:
-        await interaction.response.send_message("❌ Du hast keine Berechtigung für diesen Befehl.", ephemeral=True)
-        return
 
     cog: ChampionCog = interaction.client.get_cog("ChampionCog")
     user_id_str = str(user.id)
